test/nvme: recreate spdk/nvmecli git index file
spdk/nvmecli git index happens to be corrupted and
results in a lot of unrelated failures in the CI.
This patch attempts only to mitigate the issue of
unwanted failures by restoring nvmecli repository
to clean state.

Fixes #1496

Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/3415
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>

+++ test/common/autotest_common.sh
@@ -848,7 +848,12 @@ function nvme_cli_build() {
 	# Build against the version of SPDK under test
 	pushd $spdk_nvme_cli
 
-	git clean -dfx
+	# Remove and recreate git index in case it became corrupted
+	if ! git clean -dfx; then
+		rm -f .git/index
+		git clean -dfx
+		git reset --hard
+	fi
 
 	rm -f "$spdk_nvme_cli/spdk"
 	ln -sf "$rootdir" "$spdk_nvme_cli/spdk"
